connect store admin

+++ src/main/java/com/xpay/pay/model/Order.java
@@ -10,7 +10,7 @@ import com.xpay.pay.proxy.PaymentResponse.OrderStatus;
 public class Order {
 	private long id;
 	private String orderNo;
-	private int appId;
+	private Long appId;
 	private long storeId;
 	private StoreChannel storeChannel;
 	private long storeChannelId;
@@ -48,10 +48,10 @@ public class Order {
 	public void setOrderNo(String orderNo) {
 		this.orderNo = orderNo;
 	}
-	public int getAppId() {
+	public Long getAppId() {
 		return appId;
 	}
-	public void setAppId(int appId) {
+	public void setAppId(Long appId) {
 		this.appId = appId;
 	}
 	public long getStoreId() {
